add panel suggestion

+++ _customer/edit.php
@@ -15,21 +15,21 @@ $squery =  mysqli_query($conn, "SELECT * from customer where id = $id");
 </div>
 <div class="content">
 <a href="./" class="btn btn-secondary">Back</a>
-<form class="row g-3" method="POST" action="update.php?id=<?php echo $row['id'] ?>">
+<form id="myForm" class="row g-3" method="POST" action="update.php?id=<?php echo $row['id'] ?>">
   <h1>Customer Information</h1>
 
   <div class="col-md-3">
-    <label for="first_name" class="form-label">First Name</label>
+    <label for="first_name" class="form-label">First Name<span style="color: red;">*</span></label>
     <input required type="text" class="form-control" name="first_name" id="first_name" value="<?php echo $row['first_name'] ?>">
   </div>
 
   <div class="col-md-3">
     <label for="middle_name" class="form-label">Middle Name</label>
-    <input required type="text" class="form-control" name="middle_name" id="middle_name" value="<?php echo $row['middle_name'] ?>">
+    <input type="text" class="form-control" name="middle_name" id="middle_name" value="<?php echo $row['middle_name'] ?>">
   </div>
 
   <div class="col-md-3">
-    <label for="last_name" class="form-label">Last Name</label>
+    <label for="last_name" class="form-label">Last Name<span style="color: red;">*</span></label>
     <input required type="text" class="form-control" name="last_name" id="last_name" value="<?php echo $row['last_name'] ?>">
   </div>
 
@@ -39,7 +39,7 @@ $squery =  mysqli_query($conn, "SELECT * from customer where id = $id");
   </div>
 
   <div class="col-md-3">
-    <label for="gender" class="form-label">Gender</label>
+    <label for="gender" class="form-label">Gender<span style="color: red;">*</span></label>
     <select required id="inputState" class="form-select" name="gender">
       <option value="<?php echo $row['gender'] ?>" select hidden><?php echo $row['gender'] ?></option>
       <option value="Female">Female</option>
@@ -48,12 +48,12 @@ $squery =  mysqli_query($conn, "SELECT * from customer where id = $id");
   </div>
 
   <div class="col-md-3">
-    <label for="date_of_birth" class="form-label">Date of Birth</label>
+    <label for="date_of_birth" class="form-label">Date of Birth<span style="color: red;">*</span></label>
     <input required type="date" class="form-control" name="date_of_birth" id="date_of_birth" value="<?php echo $row['date_of_birth'] ?>">
   </div>
 
   <div class="col-md-3">
-    <label for="purok" class="form-label">Purok</label>
+    <label for="purok" class="form-label">Purok<span style="color: red;">*</span></label>
     <select required class="form-select" name="purok" id="purok">
       <option value="<?php echo $row['purok'] ?>" select hidden><?php echo $row['purok'] ?></option>
       <option value="Purok Sambag">Purok Sambag</option>
@@ -90,12 +90,12 @@ $squery =  mysqli_query($conn, "SELECT * from customer where id = $id");
   </div>
 
   <div class="col-md-3">
-    <label for="place_of_birth" class="form-label">Place of Birth</label>
+    <label for="place_of_birth" class="form-label">Place of Birth<span style="color: red;">*</span></label>
     <input required type="text" class="form-control" name="place_of_birth" id="place_of_birth"  value="<?php echo $row['place_of_birth'] ?>">
   </div>
 
   <div class="col-md-3">
-    <label for="civil_status" class="form-label">Civil Status</label>
+    <label for="civil_status" class="form-label">Civil Status<span style="color: red;">*</span></label>
     <select required class="form-select" name="civil_status" id="civil_status">
       <option value="<?php echo $row['civil_status'] ?>" select hidden><?php echo $row['civil_status'] ?></option>
       <option value="Single">Single</option>
@@ -106,14 +106,14 @@ $squery =  mysqli_query($conn, "SELECT * from customer where id = $id");
   </div>
 
   <div class="col-md-3">
-    <label for="phone_number" class="form-label">Phone Number</label>
+    <label for="phone_number" class="form-label">Phone Number<span style="color: red;">*</span></label>
     <input required type="text" class="form-control" name="phone_number" id="phone_number" value="<?php echo $row['phone_number'] ?>">
   </div>
 
   <h1>Billing Information</h1>
 
   <div class="col-md-3">
-    <label for="category" class="form-label">Category</label>
+    <label for="category" class="form-label">Category<span style="color: red;">*</span></label>
     <select required class="form-select" name="category" id="category">
     <option value="<?php echo $row['category'] ?>" select hidden><?php echo $row['category'] ?></option>
     <?php
@@ -125,19 +125,29 @@ $squery =  mysqli_query($conn, "SELECT * from customer where id = $id");
   </div>
 
   <div class="col-md-3">
-    <label for="latest_reading_date" class="form-label">Date of Latest Reading</label>
+    <label for="latest_reading_date" class="form-label">Date of Latest Reading<span style="color: red;">*</span></label>
     <input required type="date" class="form-control" name="latest_reading_date" id="latest_reading_date" value="<?php echo $row['latest_reading_date']; ?>">
   </div>
 
   <div class="col-md-3">
-    <label for="water_reading" class="form-label">Current Water Reading</label>
+    <label for="water_reading" class="form-label">Current Water Reading<span style="color: red;">*</span></label>
     <input required type="text" class="form-control" name="water_reading" id="water_reading" value="<?php echo $row['water_reading'] ?>">
   </div>
 
   <div class="col-12 buttons">
-    <button type="submit" class="btn btn-primary">Save</button>
-    <a href="./delete.php?id=<?php echo $row['id']; ?>" class="btn btn-danger">Delete</a>
+    <button type="submit" class="btn btn-primary" id="submitButton">Update</button>
+    <a href="./delete.php?id=<?php echo $row['id']; ?>" id="deleteButton" class="btn btn-danger" onclick="return confirmDelete()">Delete</a>
   </div>
+
+<script>
+function confirmDelete() {
+    // Display confirmation dialog
+    var result = confirm("Are you sure you want to delete?");
+    
+    // Return true if user clicks "OK", else return false
+    return result;
+}
+</script>
 </form>
     </div>
     <?php } ?>
